feat: extract fixRequestBody to external handler
