+++ src/handlers/public.ts
@@ -1,1 +1,2 @@
 export { responseInterceptor } from './response-interceptor';
+export { fixRequestBody } from './fix-request-body';
