+++ src/http-proxy-middleware.ts
@@ -1,10 +1,8 @@
-import { ClientRequest } from 'http';
 import type * as https from 'https';
 import type * as express from 'express';
 import type { Filter, Request, RequestHandler, Response, Options } from './types';
 import * as httpProxy from 'http-proxy';
 import { createConfig, Config } from './config-factory';
-import * as querystring from 'querystring';
 import * as contextMatcher from './context-matcher';
 import * as handlers from './_handlers';
 import { getArrow, getInstance } from './logger';
@@ -34,9 +32,6 @@ export class HttpProxyMiddleware {
 
     // log errors for debug purpose
     this.proxy.on('error', this.logError);
-
-    // fix proxied body if bodyParser is involved
-    this.proxy.on('proxyReq', this.fixBody);
 
     // https://github.com/chimurai/http-proxy-middleware/issues/19
     // expose function to upgrade externally
@@ -197,25 +192,5 @@ export class HttpProxyMiddleware {
     const errReference = 'https://nodejs.org/api/errors.html#errors_common_system_errors'; // link to Node Common Systems Errors page
 
     this.logger.error(errorMessage, requestHref, targetHref, err.code || err, errReference);
-  };
-
-  private fixBody = (proxyReq: ClientRequest, req: Request) => {
-    if (!req.body || !Object.keys(req.body).length) {
-      return;
-    }
-
-    const contentType = proxyReq.getHeader('Content-Type') as string;
-    const writeBody = (bodyData: string) => {
-      proxyReq.setHeader('Content-Length', Buffer.byteLength(bodyData));
-      proxyReq.write(bodyData);
-    };
-
-    if (contentType.includes('application/json')) {
-      writeBody(JSON.stringify(req.body));
-    }
-
-    if (contentType === 'application/x-www-form-urlencoded') {
-      writeBody(querystring.stringify(req.body));
-    }
   };
 }
